Exposed GenerateMarkdown
